requirements.txt
#--- Core Backend ---
fastapi==0.95.2
uvicorn[standard]==0.29.0
#--- Firebase Integration ---
firebase-admin==6.3.0
python-dotenv==1.0.1
#--- LLM & AI Integration ---
google-generativeai==0.3.0
#--- PDF and File Handling ---
PyPDF2==3.0.1
python-multipart==0.0.9
#--- Data Processing ---
pydantic==1.10.15
pydantic-core==2.16.3
requests==2.31.0
#--- For Dev and Testing ---
pytest==8.1.1
#--- database ---
pymongo==4.5.0
dnspython==2.4.2
#Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2